Make LHC compile with GHC 6.12
The LHC compiler now builds with GHC 6.12. Since it is so closely tied
to the GHC compiler, it is unlikely to compile successfuly with GHC
6.10 after these changes. The necessary changes are described below.

1) lhc-pkg was updated.
- The source for lhc-pkg was copied over from ghc-pkg in GHC 6.12.
- Added Binary instances for InstalledPackageInfo (copied from GHC 6.12)

2) Added support for a global package database.
The new version of ghc-pkg assumes that there is a global package
database (i.e. in $PREFIX/lib) separate from the local package
database (i.e. in ~/.lhc/). This requires creating the global package
store and passing the root location of the installation to both lhc
and lhc-pkg (i.e. the -B flag for GHC). 

3) Changed lhc and lhc-pkg into wrappers.
To support hte global package database, the lhc and lhc-pkg programs
were changed to be wrapper scripts that set the -B flag (for lhs) and
the --global-conf flag for lhc-pkg before launching the actual
executables. Thus lhc was renamed to lhc-exe and lhc-pkg was renamed
to lhc-pkg-exe.

4) Setup.hs
The setup script was modified to create the wrapper packages described
above. Also, the build results from building the library are checked
to make sure that the build succeeds and errors out if not. Finally,
a hook was added so that cabal clean will clean the libraries also.

5) Libraries
The ghc-prim and base libraries were copied over from GHC 6.12. The
integer library implementation was changed to use integer-simple from
GHC (the current integer-native did not have the proper interface).
The integer-simple is registered as integer-gmp to allow for
successful compilation of the base library.