Removed special-case prim functions from HPT analysis
The heap points-to analysis was adding special-case equations for several of
the primitive functions. Some of these equations were interacting badly with
the rest of the analysis. In particular, the primitives that were allocating
stable and weak pointers (e.g. mkStablePtr#, deRefStablePtr#) were having
problems threading the "pointerness" of their results through the rest of the
analysis. The pointers allocated by these prims were not getting modeled
correctly for the rest of the analysis.

This change is a stop-gap measure to allow the analysis to complete. The
functions are moved to the unsupported column, so that the analysis treats the
result as bottom. This causes the analysis to complete, but causes the code
generation to treat these results as "unreachable" in the code generation.

Since we can not yet compile the C code, these unreachable functions are doing
no harm at this point.  The plan is to add the primitive functions back into
the mix as we gain more understanding of how they should be modeled in the heap
analysis.