Added runtime headers and modified compiler options for C back end
The current C back end was not linking all of the necessary libraries nor did
it have access to all the necessary header files to when compiling the C file
generated from the C back end.

The header files are stored in rts/include and are available at runtime in the
$prefix/share/rts/include directory. Some ghc header files were needed for the
primitive operations a functions in the ghc-prim and base libraries. These
header files were added to rts/include/ghc directory.

The gcc compilation command was changed to add these new header file
directories to the path and to add the $prefix/lib dir to the library search
path and add the needed libraries.