Add a few primops to the C back end
The C back end needed to implement a few more primitive operations to
successfully run the "hello world" and "upto" programs. The primops are used in
core generated by GHC 6.12. The new primops allow successful running of these
simple programs.

The TopHandler was also changed so that it did not install signal handlers and
use the myThreadId# primitive. This change was needed since we do not support
threads. The new runMainIO function simply runs the main action and cleans up.