Add support for the waitRead# and readWideCharOffAddr# prims
This patch adds support for reading from input file descriptors in
lhc. The waitRead# primitive is implemented in the runtime using the
select() system call. The readWideCharOffAddr# prim is implemented
as a simple index off a base address.  These primitives add support
for reading from file descriptors (such as stdin).

Unfortunately, the support is incomplete because the implementation
in the base library uses the raise# primitive to signal EOF. This
causes us to panic at that point since we do not support any of the
exception primitives. The main problem function is the hGetChar
function in GHC.IO.Handle.Text which calls the wantReadableHandle_
function in Handle.Internal and that function raises the EOF error.