Use the RTS as a library instead of ad-hoc files
This patch changes the RTS to be a statically linked library instead
of a collection of C files that are catted together along with the
generated C code.

Along with this change is a few new runtime files for dealing with
floats.  These files are needed to provide the external functions
used by the integer-simple library. Adding these allow us to compile
the BinaryTrees benchmark which is also added in this patch.